+++ src/App.js
@@ -1,5 +1,6 @@
 import { sortedLastIndex } from 'lodash';
 import React, { Component} from 'react';
+import Member from './Member';
 import './App.css';
 
 class App extends Component {
@@ -20,29 +21,6 @@ class App extends Component {
         <Member name="近藤勇" message="JSの勉強もしていきたいです" age="18" /> 
       </div>
     );
-  }
-}
-
-class Member extends Component {
-  constructor(props) {
-    super(props);
-    this.name = props.name;
-    this.message = props.message;
-    this.age = props.age;
-    this.style = {
-      border: "2px solid black",
-      margin: "3px"
-    }
-  }
-
-  render() {
-    return (
-      <div style={this.style}>
-        <p>{this.name}</p>
-        <p>{this.message}</p>
-        <p>{this.age}</p>
-      </div>
-    )
   }
 }
